Prefer direct TypeScript dependency (#1928)
Adjust getTypeScriptLocation to pick the top-most
(in terms of directory hierarchy) TypeScript dependency it
can find.
This prevents it from finding a nested transitive
TypeScript dependency that might not match the version
requirements.

The average search performance is a bit worse now, since
it's forced to do a full directory search. Before, it
might've found a match earlier, though worst-case performance
was still more or less a full search.