Reduce noise of SCA reports (#2005)
* Drop commons-compress from nodejs-utils dependencies

The plugins using nodejs-utils, for instance sonar-css, will no longer have to bundle the unused commons-compress. That decreases the size of artifact and reduces the noise with SCA audits.

* Upgrade commons-compress to 1.20

in order to fix the potential vulnerability described in https://cve.mitre.org/cgi-bin/cvename.cgi?name=2019-12402, or at least to kill the noise in SCA audits.