EslintBridgeServer restarts always fail
# Why

the ESLint bridge server might need restarts as part of a single
scan. The scanner is trying to achieve this and currently always fails.
This is the case because the script arguments passed to the Node.js
process from the Java process are duplicated on subsequent starts. The
following excerpt from a scan log nicely shows this:

```
14:08:04.672 DEBUG: Starting Node.js process to start eslint-bridge server at port 39139
14:08:04.673 DEBUG: Launching command node --max-old-space-size=8192 --max-old-space-size=8192 /usr/src/.scannerwork/.sonartmp/eslint-bridge-bundle/package/bin/server 37023 127.0.0.1 /usr/src/.scannerwork true  39139 127.0.0.1 /usr/src/.scannerwork true
14:08:11.554 DEBUG: starting eslint-bridge server at port 37023
14:08:11.673 DEBUG: eslint-bridge server is running at port 37023
14:13:04.771 ERROR: Failed to start server (300s timeout)
org.sonarsource.nodejs.NodeCommandException: Failed to start server (300s timeout)
```

Note how the `Launching command node` log line lists all Node.js and script
arguments twice. This is the result of a missing reset of the `NodeCommandBuilder`.

# What

In order to keep my changes for this bugfix as small as possible, I opted
for an API to reset the `NodeCommandBuilder` state. This API is now used
before the `EslintBridgeServerImpl` constructs the Node.js command. This
then fixes `EslintBridgeServer` restarts.

# Future Work

`NodeCommandBuilder` is currently part of the dependency injection context.
This in itself is problematic and should be reconsidered. Maybe instead of
injecting a `NodeCommandBuilder`, instead inject `NodeCommand.ProcessWrapper`
and manually create a `NodeCommandBuilder` instance whenever necessary?

# References

 - [Bug Report](https://community.sonarsource.com/t/eslint-bridge-server-restarts-fail-due-to-missing-script-argument-resets/47292)