Resolve JS-278
This commit removes the concept of ESLint key from the backend.

The rational is that the usage of ESLint is an implementation detail of the frontend that must not propagate
to the other layers: would we want to totally change the way the frontend detect issues, the backend should
not be impacted at all.

Rationality aside, this change removes the requirements of maintaining a mapping between the ESLint keys and
their RSpec counterparts, which will make our life easier in the future.

Java checks now have RSpec identifiers only. To preserve compatiblity with potential
plugins that depend on this one, the semantic `eslintKey` was left untouched, but now means
"RSpec identifier". Only the tests that contained the rule key in their expectation have been updated.
Other tests, and especially ruling tests, were left untouched.

A word about the ESLint plugin: in order to not break the public contract of the ESLint plugin that
is also built from the frontend code, the ESLint keys have been kept in the rules entry point.